unify: FP16 prune for all builds + float16_t typedef + meson gating
Store raw build vectors as float16_t (2 bytes/elem vs 4 for FP32), halving
the construct-time memory. The FP16 L2sq occlusion check replaces PQ
code_distance in ALL build paths (K=1 and K>1), unifying the prune distance
computation into one canonical path.

float16_t = __fp16 (native on ARM NEON + x86 F16C). Meson detects FP16
hardware support at configure time (no forced -mf16c flag — works with
-march=native, -mcpu=neoverse-v2, etc.).

l2sq_f16 takes float16_t* directly (no per-call conversion). NEON: vld1_f16
+ vcvt_f32_f16 + FMLA. x86: _mm256_cvtph_ps + FMA. FP32 accumulation avoids
overflow at 768-dim.

K=1: build 90s, recall 0.9940. K=4: build 78s, recall 0.9958 (was 0.9931
with PQ prune — the partitioned path now gets the same FP16 quality).
All 12 tests pass.

Microbench (pastry x86 i7-8809G): FP16+F16C is 1.87× faster than FP32 AVX2
(118ns vs 222ns per 768-dim L2sq) due to halved load bandwidth.

+++ test/test_vamana_core.cpp
@@ -181,7 +181,7 @@ TEST(VamanaCore, InsertBuildFromVector) {
     const uint16_t R = 6;
     TestGraph g(N, R, /*L=*/12);
 
-    std::vector<float> vec(g.params.dim, 0.0f);
+    std::vector<float16_t> vec(g.params.dim, static_cast<float16_t>(0));
     VamanaTLS tls;
     tls.resize(N);
     tls.resize_lut(g.quant.lut_size());
